Generate cleaner/smaller build output
The key difference is that the produced output of the build step is not
outputting files into an extra lib/ folder anymore, but instead into the
package's root.

Unfortunately, npm/yarn publish include the same package.json from the
top of the project's root into this package.

This new build process still uses the package.json as the key input for
the built output, but strips off stuff that's irrelevant to the *NPM*
package, like devDependencies, scripts, etc.